Raises NotFoundError only when status code is 404

+++ mozdownload/scraper.py
@@ -311,10 +311,13 @@ class Scraper(object):
                 if log_level <= logging.INFO and content_length:
                     pbar.finish()
 
-            except requests.exceptions.HTTPError:
-                if os.path.isfile(tmp_file):
-                    os.remove(tmp_file)
-                raise errors.NotFoundError("The requested file was not found", self.url)
+            except requests.exceptions.HTTPError as ex:
+                if ex.response.status_code == 404:
+                    if os.path.isfile(tmp_file):
+                        os.remove(tmp_file)
+                    raise errors.NotFoundError("The requested file was not found", self.url)
+                else:
+                    raise
 
             except Exception:
                 if os.path.isfile(tmp_file):
